requirements.txt
Django
djangorestframework
django-cors-headers
pyjwt
autopep8